requirements.txt
pyzk==0.9.0
tkinter-tooltip==2.0.0
Pillow>=10.1.0
psycopg2-binary==2.9.9
bcrypt==4.1.2
numpy>=1.24.0 